Restore core-executor session + queries tests
