+++ crates/core-executor/Cargo.toml
@@ -47,6 +47,7 @@ uuid = { workspace = true }
 [dev-dependencies]
 bytes = { workspace = true }
 insta = { version = "1.42.0", features = ["yaml", "filters"] }
+paste = "1"
 
 [lints]
 workspace = true
