+++ crates/core-executor/src/tests/table_setup.sql
