--- crates/runtime/src/tests/mod.rs
@@ -1,7 +1,0 @@
-mod auth;
-mod queries;
-pub mod server;
-mod session;
-pub mod utils;
-
-pub use server::*;
