--- crates/runtime/src/tests/queries.rs
@@ -1,24 +1,0 @@
-use crate::tests::utils::macros::test_query;
-
-test_query!(select_date_add_diff, "SELECT dateadd(day, 5, '2025-06-01')");
-test_query!(func_date_add, "SELECT date_add(day, 30, '2025-01-06')");
-// // SELECT
-test_query!(select_star, "SELECT * FROM employee_table");
-// FIXME: ILIKE is not supported yet
-// test_query!(select_ilike, "SELECT * ILIKE '%id%' FROM employee_table;");
-test_query!(
-    select_exclude,
-    "SELECT * EXCLUDE department_id FROM employee_table;"
-);
-test_query!(
-    select_exclude_multiple,
-    "SELECT * EXCLUDE (department_id, employee_id) FROM employee_table;"
-);
-
-test_query!(
-    qualify,
-    "SELECT product_id, retail_price, quantity, city
-    FROM sales
-    QUALIFY ROW_NUMBER() OVER (PARTITION BY city ORDER BY retail_price) = 1
-    ;"
-);
